New Features
 - Multi-channel (>2) support added. Allows you to select any number of active inputs and outputs.
 - Input Mixer added to control how the inputs are mixed and grouped for sending and for monitoring. Can now choose to send to others in Mono, Stereo, or Multichannel (all active inputs sent to others). When you are receiving multichannel from others, allows you to choose how to mix them locally, including optional independent volume/pan/mute/solo. It's a lot of new stuff here.
 - Input effects are now per channel as configured by grouping
 - Full gate/compressor/eq effects now available per user track (and even per channel there if desired)
 - You can route remote users to specific outputs (if you have activated them, or are using the plugin in multi-bus capable host)
 - Can now insert Audio Unit plugin in "instrument" slots so you can get access to multi-bus outputs on hosts like Logic...
 - Added minimal/full view mode choice... to let you hide the statistics areas of connected users for a cleaner look. You can toggle the full view for all, or individually by clicking on the user name
 - Recording filenames now include group name
 - Added support for MP3 playing on platforms that didn't have it
 - Initially mute incoming audio of new connections while they are auto-adjusting jitter buffers until stability is achieved.
 - Added Default User Volume level in options, which determines the initial volume new connecting users get when they haven't been seen before.
 - Text Chat added, for another way to communicate within a group during a  session. URLs posted in chat messages can be clicked to open the links in a browser.
 - Self-monitoring delay (and backing track delay) options to help overcome latency related rhythmic issues.
 - Added Group Latency Match feature
 - Added command line options to start sonobus and immediately connect, or to load a saved setup, along with basic "headless" operation mode for power users.
 - Initial translation support for several languages... still a work in progress.
 - Many bugfixes
 - WARNING - using Send Multichannel in a group with users of the older version of the app will crash the old application... this is unfortunate, but also a good incentive for them to  update!

+++ Builds/LinuxMakefile/uninstall.sh
@@ -8,14 +8,26 @@ fi
 
 echo "Un-Installing SonoBus from ${PREFIX} ... (specify destination as command line argument if you have it elsewhere)"
 
-if ! rm -f ${PREFIX}/bin/SonoBus ; then
-  echo
-  echo "Looks like you need to run this with 'sudo $0'"
-  exit 2
+# remove old binary name
+if [ -f ${PREFIX}/bin/SonoBus ] ; then
+  if ! rm -f ${PREFIX}/bin/SonoBus ; then
+    echo
+    echo "Looks like you need to run this with 'sudo $0'"
+    exit 2
+  fi
+fi
+
+if [ -f ${PREFIX}/bin/sonobus ] ; then
+  if ! rm -f ${PREFIX}/bin/sonobus ; then
+    echo
+    echo "Looks like you need to run this with 'sudo $0'"
+    exit 2
+  fi
 fi
 
 rm -f ${PREFIX}/share/applications/sonobus.desktop
 rm -f ${PREFIX}/pixmaps/sonobus.png
 rm -rf ${PREFIX}/lib/vst3/SonoBus.vst3
+rm -rf ${PREFIX}/lib/vst3/sonobus.vst3
 
 echo "SonoBus uninstalled"
